mcp-tools/chrome-auth/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chrome-auth"
version = "0.1.0"
description = "Chrome authentication provider — extract cookies, localStorage, and request headers from Chrome via cookie DB or CDP."
requires-python = ">=3.9"
dependencies = [
    "browser-cookie3>=0.19.1",
    "websocket-client>=1.6.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["pytest"]
